requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Machine Learning
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Dashboard
streamlit==1.28.1
streamlit-option-menu==0.3.6

# MLOps
mlflow==2.8.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing
openpyxl==3.1.2
python-dateutil==2.8.2

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
faker==19.13.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0

# Jupyter
jupyter==1.0.0
ipykernel==6.26.0
notebook==7.0.6

# File handling
PyYAML==6.0.1